requirements.txt
streamlit
langchain
langchain-community
chromadb
python-dotenv

